+++ packages/tokens/scripts/.eslintrc
