+++ packages/tokens/scripts/outputs/scss.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 import fs from 'fs';
 import { pascalCase } from 'change-case';
 import chalk from 'chalk';
@@ -7,7 +6,7 @@ import { createFolder, createHeaderComment } from '../helpers.js';
 /**
  * Creates scss exports from the provided css file
  *
- * @param {string} header The header to prepend to the output
+ * @param {string[]} header The header to prepend to the output
  * @param {string} inputFile Path to the input file
  * @param {string} outputFile Path to the output file
  */
