+++ packages/tokens/scripts/transforms/index.js
@@ -1,5 +1,5 @@
-export * from './add-color-name.js';
+export * from './add-color-prefix.js';
 export * from './add-fallback-fonts.js';
-export * from './add-missing-quotes-for-strings.js';
-export * from './calc.js';
-export * from './log.js';
+export * from './add-missing-quotes.js';
+export * from './adjust-shadow.js';
+export * from './use-css-calc.js';
