--- packages/tokens/scripts/transforms/log.js
@@ -1,9 +1,0 @@
-/**
- * This transformer is only used for easier debugging and can be activated when needed
- */
-export const log = {
-  matcher: () => true,
-  name: 'syn/log',
-  transformer: (token) => token.value,
-  type: 'value',
-};
