--- packages/tokens2/.eslintrc.json
@@ -1,26 +1,0 @@
-{
-  "ignorePatterns": [
-    "node_modules/"
-  ],
-  "extends": [
-    "@synergy-design-system/eslint-config-syn/ts"
-  ],
-  "parserOptions": {
-    "project": "./tsconfig.lint.json"
-  },
-  "rules": {
-    "no-param-reassign": "off"
-  },
-  "settings": {
-    "import/resolver": {
-      "node": {
-        "extensions": [".ts", ".js"],
-        "moduleDirectory": ["node_modules", "scripts", "dist"]
-      },
-      "typescript": {
-        "extensions": [".ts", ".js"],
-        "moduleDirectory": ["node_modules", "scripts", "dist"]
-      }
-    }
-  }
-}
