--- packages/tokens2/.stylelintrc.json
@@ -1,10 +1,0 @@
-{
-  "extends": "@synergy-design-system/stylelint-config-syn",
-  "rules": {
-    "alpha-value-notation": null,
-    "color-hex-length": null,
-    "color-function-notation": null,
-    "value-keyword-case": null,
-    "order/order": null
-  }
-}
