--- packages/tokens2/package.json
@@ -1,117 +1,0 @@
-{
-  "author": {
-    "name": "SICK Global UX Foundation",
-    "url": "https://www.sick.com"
-  },
-  "description": "Design tokens for Synergy, the SICK Design System",
-  "devDependencies": {
-    "@synergy-design-system/eslint-config-syn": "workspace:*",
-    "@synergy-design-system/stylelint-config-syn": "workspace:*",
-    "@tokens-studio/sd-transforms": "^1.2.8",
-    "@types/node": "^20.17.6",
-    "chalk": "^5.3.0",
-    "change-case": "^5.4.4",
-    "eslint": "^8.57.1",
-    "eslint-import-resolver-typescript": "^3.6.3",
-    "rimraf": "^6.0.1",
-    "style-dictionary": "^4.2.0",
-    "stylelint": "^16.10.0",
-    "typescript": "^5.7.2"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/js/index.d.ts",
-      "import": "./dist/js/index.js",
-      "default": "./dist/js/index.js"
-    },
-    "./package.json": {
-      "default": "./package.json"
-    },
-    "./themes/*": "./dist/themes/*",
-    "./scss/*": "./dist/scss/*"
-  },
-  "files": [
-    "dist",
-    "CHANGELOG.md",
-    "src/figma-tokens"
-  ],
-  "homepage": "https://synergy-design-system.github.io",
-  "keywords": [
-    "SDS",
-    "SICK",
-    "Synergy Design System",
-    "Design Tokens"
-  ],
-  "license": "MIT",
-  "name": "@synergy-design-system/tokens",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/synergy-design-system/synergy-design-system.git",
-    "directory": "packages/tokens"
-  },
-  "release": {
-    "branches": [
-      "main",
-      {
-        "name": "beta",
-        "prerelease": true
-      },
-      {
-        "name": "alpha",
-        "prerelease": true
-      }
-    ],
-    "plugins": [
-      "@semantic-release/commit-analyzer",
-      "@semantic-release/release-notes-generator",
-      [
-        "@semantic-release/changelog",
-        {
-          "changelogFile": "CHANGELOG.md"
-        }
-      ],
-      [
-        "@semantic-release/exec",
-        {
-          "prepareCmd": "pnpm build"
-        }
-      ],
-      [
-        "@semantic-release/npm",
-        {
-          "npmPublish": false
-        }
-      ],
-      [
-        "@semantic-release/git",
-        {
-          "message": "chore(release/tokens): ${nextRelease.version} [skip actions]\n\n${nextRelease.notes}",
-          "assets": [
-            "CHANGELOG.md",
-            "package.json"
-          ]
-        }
-      ],
-      [
-        "@semantic-release/github",
-        {
-          "successComment": false
-        }
-      ]
-    ]
-  },
-  "scripts": {
-    "build": "pnpm clean && node scripts/build.js",
-    "clean": "rimraf build",
-    "lint:css": "stylelint --allow-empty-input \"dist/**/*.css\"",
-    "lint:js": "eslint scripts dist/js",
-    "lint:types": "tsc --noEmit",
-    "lint": "pnpm run /^lint:.*/",
-    "start": "pnpm build",
-    "release": "semantic-release  --tagFormat 'tokens/${version}' -e semantic-release-monorepo",
-    "release.dry": "semantic-release --dry-run --tagFormat 'tokens/${version}' -e semantic-release-monorepo"
-  },
-  "type": "module",
-  "types": "./dist/js/index.d.ts",
-  "version": "2.15.0"
-}
