--- packages/tokens2/scripts/formats/index.js
@@ -1,1 +1,0 @@
-export * from './css-vars.js';
