--- packages/tokens2/scripts/outputs/index.js
@@ -1,2 +1,0 @@
-export * from './js.js';
-export * from './scss.js';
