--- packages/tokens2/scripts/transforms/index.js
@@ -1,5 +1,0 @@
-export * from './add-color-prefix.js';
-export * from './add-fallback-fonts.js';
-export * from './add-missing-quotes.js';
-export * from './adjust-shadow.js';
-export * from './use-css-calc.js';
