--- packages/tokens2/src/figma-tokens/$metadata.json
@@ -1,18 +1,0 @@
-{
-  "tokenSetOrder": [
-    "globals",
-    "color/dark",
-    "color/light",
-    "color/primitives",
-    "semantic/focus-ring",
-    "semantic/input",
-    "semantic/button",
-    "semantic/toggle",
-    "semantic/typography",
-    "semantic/link",
-    "semantic/overlays",
-    "semantic/panels",
-    "semantic/tooltip",
-    "_docs"
-  ]
-}
