--- packages/tokens2/src/figma-tokens/semantic/button.json
@@ -1,18 +1,0 @@
-{
-  "button": {
-    "font-size": {
-      "small": {
-        "value": "{font-size.small}",
-        "type": "fontSizes"
-      },
-      "medium": {
-        "value": "{font-size.medium}",
-        "type": "fontSizes"
-      },
-      "large": {
-        "value": "{font-size.large}",
-        "type": "fontSizes"
-      }
-    }
-  }
-}
