--- packages/tokens2/src/figma-tokens/semantic/focus-ring.json
@@ -1,16 +1,0 @@
-{
-  "focus-ring": {
-    "color": {
-      "value": "{primary.400}",
-      "type": "color"
-    },
-    "width": {
-      "value": "{border-width.medium}",
-      "type": "borderWidth"
-    },
-    "offset": {
-      "value": "{spacing.3x-small}",
-      "type": "spacing"
-    }
-  }
-}
