--- packages/tokens2/src/figma-tokens/semantic/link.json
@@ -1,30 +1,0 @@
-{
-  "link": {
-    "color": {
-      "value": "{primary.600}",
-      "type": "color"
-    },
-    "color-hover": {
-      "value": "{primary.900}",
-      "type": "color"
-    },
-    "color-active": {
-      "value": "{primary.950}",
-      "type": "color"
-    },
-    "quiet": {
-      "color": {
-        "value": "{typography.color.text}",
-        "type": "color"
-      },
-      "color-hover": {
-        "value": "{primary.900}",
-        "type": "color"
-      },
-      "color-active": {
-        "value": "{primary.950}",
-        "type": "color"
-      }
-    }
-  }
-}
