--- packages/tokens2/src/figma-tokens/semantic/overlays.json
@@ -1,10 +1,0 @@
-{
-  "overlay": {
-    "background": {
-      "color": {
-        "value": "rgba(0, 0, 0, 0.5)",
-        "type": "color"
-      }
-    }
-  }
-}
