--- packages/tokens2/src/figma-tokens/semantic/panels.json
@@ -1,20 +1,0 @@
-{
-  "panel": {
-    "background": {
-      "color": {
-        "value": "{neutral.0}",
-        "type": "color"
-      }
-    },
-    "border": {
-      "color": {
-        "value": "{neutral.300}",
-        "type": "color"
-      },
-      "width": {
-        "value": "{border-width.small}",
-        "type": "borderWidth"
-      }
-    }
-  }
-}
