--- packages/tokens2/src/figma-tokens/semantic/toggle.json
@@ -1,18 +1,0 @@
-{
-  "toggle": {
-    "size": {
-      "small": {
-        "value": "14px",
-        "type": "spacing"
-      },
-      "medium": {
-        "value": "{spacing.medium}",
-        "type": "spacing"
-      },
-      "large": {
-        "value": "{spacing.large}",
-        "type": "spacing"
-      }
-    }
-  }
-}
